offscreen-canvas/the-offscreen-canvas/size.attributes.parse.empty.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>OffscreenCanvas test: size.attributes.parse.empty</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/canvas-tests.js"></script>

<h1>size.attributes.parse.empty</h1>
<p class="desc">Parsing of non-negative integers</p>


<script>
var t = async_test("Parsing of non-negative integers");
t.step(function() {

var offscreenCanvas = new OffscreenCanvas(100, 50);
var ctx = offscreenCanvas.getContext('2d');

_assertSame(offscreenCanvas.width, 300, "offscreenCanvas.width", "300");
_assertSame(offscreenCanvas.height, 150, "offscreenCanvas.height", "150");

t.done();

});
</script>
